requirements.txt
azure-functions==1.21.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
slack_sdk==3.34.0
urllib3==2.3.0
